Refactor code: move sort utils under logic
