+++ src/main/java/seedu/fast/logic/parser/SortCommandParser.java
@@ -2,9 +2,9 @@ package seedu.fast.logic.parser;
 
 import static seedu.fast.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
-import seedu.fast.commons.util.sort.SortByAppointment;
-import seedu.fast.commons.util.sort.SortByName;
-import seedu.fast.commons.util.sort.SortByPriority;
+import seedu.fast.logic.parser.sort.SortByAppointment;
+import seedu.fast.logic.parser.sort.SortByName;
+import seedu.fast.logic.parser.sort.SortByPriority;
 import seedu.fast.logic.commands.SortCommand;
 import seedu.fast.logic.parser.exceptions.ParseException;
