+++ src/main/java/seedu/fast/logic/parser/sort/SortByAppointment.java
@@ -1,19 +1,19 @@
-package seedu.fast.commons.util.sort;
+package seedu.fast.logic.parser.sort;
 
 import java.util.Comparator;
 
 import seedu.fast.model.person.Person;
 
 /**
- * Custom comparator that compares Person objects by their Names.
+ * Custom comparator that compares Person objects by their Appointment date.
  */
-public class SortByName implements Comparator<Person> {
+public class SortByAppointment implements Comparator<Person> {
 
-    public static final String KEYWORD = "name";
+    public static final String KEYWORD = "appointment";
 
     @Override
     public int compare(Person p1, Person p2) {
-        return p1.getName().fullName.compareTo(p2.getName().fullName);
+        return p1.getAppointment().convertDate().compareTo(p2.getAppointment().convertDate());
     }
 
     @Override
@@ -24,7 +24,7 @@ public class SortByName implements Comparator<Person> {
         }
 
         // instanceof handles nulls
-        if (other instanceof SortByName) {
+        if (other instanceof SortByAppointment) {
             return true;
         }
