+++ src/main/java/seedu/fast/logic/parser/sort/SortByPriority.java
@@ -1,19 +1,19 @@
-package seedu.fast.commons.util.sort;
+package seedu.fast.logic.parser.sort;
 
 import java.util.Comparator;
 
 import seedu.fast.model.person.Person;
 
 /**
- * Custom comparator that compares Person objects by their Appointment date.
+ * Custom comparator that compares Person objects by their Tags.
  */
-public class SortByAppointment implements Comparator<Person> {
+public class SortByPriority implements Comparator<Person> {
 
-    public static final String KEYWORD = "appointment";
+    public static final String KEYWORD = "priority";
 
     @Override
     public int compare(Person p1, Person p2) {
-        return p1.getAppointment().convertDate().compareTo(p2.getAppointment().convertDate());
+        return p1.getPriority() - p2.getPriority();
     }
 
     @Override
@@ -24,7 +24,7 @@ public class SortByAppointment implements Comparator<Person> {
         }
 
         // instanceof handles nulls
-        if (other instanceof SortByAppointment) {
+        if (other instanceof SortByPriority) {
             return true;
         }
